+++ homeassistant/components/moat/sensor.py
@@ -7,99 +7,100 @@ from moat_ble import DeviceClass, DeviceKey, SensorUpdate, Units
 from homeassistant import config_entries
 from homeassistant.components.bluetooth.passive_update_processor import (
     PassiveBluetoothDataProcessor,
-    PassiveBluetoothDataUpdate,
-    PassiveBluetoothEntityKey,
+    # PassiveBluetoothDataUpdate,
+    # PassiveBluetoothEntityKey,
     PassiveBluetoothProcessorCoordinator,
     PassiveBluetoothProcessorEntity,
 )
 from homeassistant.components.sensor import (
     SensorDeviceClass,
     SensorEntity,
-    SensorEntityDescription,
-    SensorStateClass,
+    # SensorEntityDescription,
+    # SensorStateClass,
 )
-from homeassistant.const import (
-    PERCENTAGE,
-    SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
-    UnitOfElectricPotential,
-    UnitOfTemperature,
-)
+
+# from homeassistant.const import (
+#     PERCENTAGE,
+#     SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
+#     UnitOfElectricPotential,
+#     UnitOfTemperature,
+# )
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddConfigEntryEntitiesCallback
-from homeassistant.helpers.sensor import sensor_device_info_to_hass_device_info
+from homeassistant.helpers.sensor import sensor_update_to_bluetooth_data_update
 
 from .const import DOMAIN
 
-SENSOR_DESCRIPTIONS = {
-    (DeviceClass.TEMPERATURE, Units.TEMP_CELSIUS): SensorEntityDescription(
-        key=f"{DeviceClass.TEMPERATURE}_{Units.TEMP_CELSIUS}",
-        device_class=SensorDeviceClass.TEMPERATURE,
-        native_unit_of_measurement=UnitOfTemperature.CELSIUS,
-        state_class=SensorStateClass.MEASUREMENT,
-    ),
-    (DeviceClass.HUMIDITY, Units.PERCENTAGE): SensorEntityDescription(
-        key=f"{DeviceClass.HUMIDITY}_{Units.PERCENTAGE}",
-        device_class=SensorDeviceClass.HUMIDITY,
-        native_unit_of_measurement=PERCENTAGE,
-        state_class=SensorStateClass.MEASUREMENT,
-    ),
-    (DeviceClass.BATTERY, Units.PERCENTAGE): SensorEntityDescription(
-        key=f"{DeviceClass.BATTERY}_{Units.PERCENTAGE}",
-        device_class=SensorDeviceClass.BATTERY,
-        native_unit_of_measurement=PERCENTAGE,
-        state_class=SensorStateClass.MEASUREMENT,
-    ),
-    (DeviceClass.VOLTAGE, Units.ELECTRIC_POTENTIAL_VOLT): SensorEntityDescription(
-        key=f"{DeviceClass.VOLTAGE}_{Units.ELECTRIC_POTENTIAL_VOLT}",
-        device_class=SensorDeviceClass.VOLTAGE,
-        native_unit_of_measurement=UnitOfElectricPotential.VOLT,
-        state_class=SensorStateClass.MEASUREMENT,
-    ),
-    (
-        DeviceClass.SIGNAL_STRENGTH,
-        Units.SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
-    ): SensorEntityDescription(
-        key=f"{DeviceClass.SIGNAL_STRENGTH}_{Units.SIGNAL_STRENGTH_DECIBELS_MILLIWATT}",
-        device_class=SensorDeviceClass.SIGNAL_STRENGTH,
-        native_unit_of_measurement=SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
-        state_class=SensorStateClass.MEASUREMENT,
-        entity_registry_enabled_default=False,
-    ),
-}
+# SENSOR_DESCRIPTIONS = {
+#     (DeviceClass.TEMPERATURE, Units.TEMP_CELSIUS): SensorEntityDescription(
+#         key=f"{DeviceClass.TEMPERATURE}_{Units.TEMP_CELSIUS}",
+#         device_class=SensorDeviceClass.TEMPERATURE,
+#         native_unit_of_measurement=UnitOfTemperature.CELSIUS,
+#         state_class=SensorStateClass.MEASUREMENT,
+#     ),
+#     (DeviceClass.HUMIDITY, Units.PERCENTAGE): SensorEntityDescription(
+#         key=f"{DeviceClass.HUMIDITY}_{Units.PERCENTAGE}",
+#         device_class=SensorDeviceClass.HUMIDITY,
+#         native_unit_of_measurement=PERCENTAGE,
+#         state_class=SensorStateClass.MEASUREMENT,
+#     ),
+#     (DeviceClass.BATTERY, Units.PERCENTAGE): SensorEntityDescription(
+#         key=f"{DeviceClass.BATTERY}_{Units.PERCENTAGE}",
+#         device_class=SensorDeviceClass.BATTERY,
+#         native_unit_of_measurement=PERCENTAGE,
+#         state_class=SensorStateClass.MEASUREMENT,
+#     ),
+#     (DeviceClass.VOLTAGE, Units.ELECTRIC_POTENTIAL_VOLT): SensorEntityDescription(
+#         key=f"{DeviceClass.VOLTAGE}_{Units.ELECTRIC_POTENTIAL_VOLT}",
+#         device_class=SensorDeviceClass.VOLTAGE,
+#         native_unit_of_measurement=UnitOfElectricPotential.VOLT,
+#         state_class=SensorStateClass.MEASUREMENT,
+#     ),
+#     (
+#         DeviceClass.SIGNAL_STRENGTH,
+#         Units.SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
+#     ): SensorEntityDescription(
+#         key=f"{DeviceClass.SIGNAL_STRENGTH}_{Units.SIGNAL_STRENGTH_DECIBELS_MILLIWATT}",
+#         device_class=SensorDeviceClass.SIGNAL_STRENGTH,
+#         native_unit_of_measurement=SIGNAL_STRENGTH_DECIBELS_MILLIWATT,
+#         state_class=SensorStateClass.MEASUREMENT,
+#         entity_registry_enabled_default=False,
+#     ),
+# }
 
 
-def _device_key_to_bluetooth_entity_key(
-    device_key: DeviceKey,
-) -> PassiveBluetoothEntityKey:
-    """Convert a device key to an entity key."""
-    return PassiveBluetoothEntityKey(device_key.key, device_key.device_id)
-
-
-def sensor_update_to_bluetooth_data_update(
-    sensor_update: SensorUpdate,
-) -> PassiveBluetoothDataUpdate:
-    """Convert a sensor update to a bluetooth data update."""
-    return PassiveBluetoothDataUpdate(
-        devices={
-            device_id: sensor_device_info_to_hass_device_info(device_info)
-            for device_id, device_info in sensor_update.devices.items()
-        },
-        entity_descriptions={
-            _device_key_to_bluetooth_entity_key(device_key): SENSOR_DESCRIPTIONS[
-                (description.device_class, description.native_unit_of_measurement)
-            ]
-            for device_key, description in sensor_update.entity_descriptions.items()
-            if description.device_class and description.native_unit_of_measurement
-        },
-        entity_data={
-            _device_key_to_bluetooth_entity_key(device_key): sensor_values.native_value
-            for device_key, sensor_values in sensor_update.entity_values.items()
-        },
-        entity_names={
-            _device_key_to_bluetooth_entity_key(device_key): sensor_values.name
-            for device_key, sensor_values in sensor_update.entity_values.items()
-        },
-    )
+# def _device_key_to_bluetooth_entity_key(
+#     device_key: DeviceKey,
+# ) -> PassiveBluetoothEntityKey:
+#     """Convert a device key to an entity key."""
+#     return PassiveBluetoothEntityKey(device_key.key, device_key.device_id)
+#
+#
+# def sensor_update_to_bluetooth_data_update(
+#     sensor_update: SensorUpdate,
+# ) -> PassiveBluetoothDataUpdate:
+#     """Convert a sensor update to a bluetooth data update."""
+#     return PassiveBluetoothDataUpdate(
+#         devices={
+#             device_id: sensor_device_info_to_hass_device_info(device_info)
+#             for device_id, device_info in sensor_update.devices.items()
+#         },
+#         entity_descriptions={
+#             _device_key_to_bluetooth_entity_key(device_key): SENSOR_DESCRIPTIONS[
+#                 (description.device_class, description.native_unit_of_measurement)
+#             ]
+#             for device_key, description in sensor_update.entity_descriptions.items()
+#             if description.device_class and description.native_unit_of_measurement
+#         },
+#         entity_data={
+#             _device_key_to_bluetooth_entity_key(device_key): sensor_values.native_value
+#             for device_key, sensor_values in sensor_update.entity_values.items()
+#         },
+#         entity_names={
+#             _device_key_to_bluetooth_entity_key(device_key): sensor_values.name
+#             for device_key, sensor_values in sensor_update.entity_values.items()
+#         },
+#     )
 
 
 async def async_setup_entry(
